refactor: restructured index files into index folder
