+++ app/index/index.html
@@ -16,9 +16,9 @@
       <img id="logo" src="./images/eyeslogo128.png"></img>
     </div>
 
-    <script src="./indexController.js"></script>
-    <script src="./indexModel.js"></script>
-    <script src="./indexView.js"></script>
-    <script src="./injectHTML.js"></script>
+    <script src="./index/indexController.js"></script>
+    <script src="./index/indexModel.js"></script>
+    <script src="./index/indexView.js"></script>
+    <script src="./index/injectHTML.js"></script>
   </body>
 </html>
